requirements.txt
-i https://pypi.org/simple
altair==5.5.0; python_version >= '3.9'
anyio==4.7.0; python_version >= '3.9'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
arrow==1.3.0; python_version >= '3.8'
asttokens==3.0.0; python_version >= '3.8'
async-lru==2.0.4; python_version >= '3.8'
attrs==24.2.0; python_version >= '3.7'
babel==2.16.0; python_version >= '3.8'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
bleach==6.2.0; python_version >= '3.9'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.12.14; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
comm==0.2.2; python_version >= '3.8'
contourpy==1.3.1; python_version >= '3.10'
cycler==0.12.1; python_version >= '3.8'
debugpy==1.8.11; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
executing==2.1.0; python_version >= '3.8'
fastjsonschema==2.21.1
feature-engine==1.8.2; python_full_version >= '3.9.0'
filelock==3.16.1; python_version >= '3.8'
flask==3.1.0; python_version >= '3.9'
fonttools==4.55.3; python_version >= '3.8'
fqdn==1.5.1
fsspec==2024.10.0; python_version >= '3.8'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.43; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
ipykernel==6.29.5; python_version >= '3.8'
ipython==8.30.0; python_version >= '3.10'
ipywidgets==8.1.5; python_version >= '3.7'
isoduration==20.11.0
itsdangerous==2.2.0; python_version >= '3.8'
jedi==0.19.2; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
json5==0.10.0; python_full_version >= '3.8.0'
jsonpointer==3.0.0
jsonschema[format-nongpl]==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
jupyter==1.0.0
jupyter-client==8.6.3; python_version >= '3.8'
jupyter-console==6.6.3; python_version >= '3.7'
jupyter-core==5.7.2; python_version >= '3.8'
jupyter-events==0.10.0; python_version >= '3.8'
jupyter-lsp==2.2.5; python_version >= '3.8'
jupyter-server==2.14.2; python_version >= '3.8'
jupyter-server-terminals==0.5.3; python_version >= '3.8'
jupyterlab==4.2.6; python_version >= '3.8'
jupyterlab-pygments==0.3.0; python_version >= '3.8'
jupyterlab-server==2.27.3; python_version >= '3.8'
jupyterlab-widgets==3.0.13; python_version >= '3.7'
kaggle==1.6.17
kiwisolver==1.4.7; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
matplotlib==3.9.2; python_version >= '3.9'
matplotlib-inline==0.1.7; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
mistune==3.0.2; python_version >= '3.7'
mpmath==1.3.0
narwhals==1.18.4; python_version >= '3.8'
nbclient==0.10.1; python_full_version >= '3.8.0'
nbconvert==7.16.4; python_version >= '3.8'
nbformat==5.10.4; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
networkx==3.4.2; python_version >= '3.10'
notebook==7.2.1; python_version >= '3.8'
notebook-shim==0.2.4; python_version >= '3.7'
numpy==2.1.3; python_version >= '3.10'
nvidia-cublas-cu12==12.4.5.8; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-cupti-cu12==12.4.127; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-nvrtc-cu12==12.4.127; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-runtime-cu12==12.4.127; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cudnn-cu12==9.1.0.70; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cufft-cu12==11.2.1.3; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-curand-cu12==10.3.5.147; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusolver-cu12==11.6.1.9; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusparse-cu12==12.3.1.170; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nccl-cu12==2.21.5; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nvjitlink-cu12==12.4.127; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nvtx-cu12==12.4.127; platform_system == 'Linux' and platform_machine == 'x86_64'
opendatasets==0.1.22
overrides==7.7.0; python_version >= '3.6'
packaging==24.2; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
pandocfilters==1.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.4; python_version >= '3.6'
patsy==1.0.1; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
pillow==11.0.0; python_version >= '3.9'
platformdirs==4.3.6; python_version >= '3.8'
prometheus-client==0.21.1; python_version >= '3.8'
prompt-toolkit==3.0.48; python_full_version >= '3.7.0'
protobuf==5.29.1; python_version >= '3.8'
psutil==6.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ptyprocess==0.7.0; os_name != 'nt'
pure-eval==0.2.3
pyarrow==18.1.0; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
pydeck==0.9.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyparsing==3.2.0; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-json-logger==3.2.0; python_version >= '3.8'
python-slugify==8.0.4; python_version >= '3.7'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
pyzmq==26.2.0; python_version >= '3.7'
qtconsole==5.6.1; python_version >= '3.8'
qtpy==2.4.2; python_version >= '3.7'
referencing==0.35.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rfc3986-validator==0.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rich==13.9.4; python_full_version >= '3.8.0'
rpds-py==0.22.3; python_version >= '3.9'
scikit-learn==1.5.1; python_version >= '3.9'
scipy==1.14.1; python_version >= '3.10'
seaborn==0.13.2; python_version >= '3.8'
send2trash==1.8.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
setuptools==71.1.0; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
stack-data==0.6.3
statsmodels==0.14.4; python_version >= '3.9'
streamlit==1.40.2; python_version >= '3.9' and python_full_version != '3.9.7'
sympy==1.13.1; python_version >= '3.9'
tenacity==9.0.0; python_version >= '3.8'
terminado==0.18.1; python_version >= '3.8'
text-unidecode==1.3
threadpoolctl==3.5.0; python_version >= '3.8'
tinycss2==1.4.0; python_version >= '3.8'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
torch==2.5.1; python_full_version >= '3.8.0'
tornado==6.4.2; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
traitlets==5.14.3; python_version >= '3.8'
triton==3.1.0; python_version < '3.13' and platform_system == 'Linux' and platform_machine == 'x86_64'
types-python-dateutil==2.9.0.20241206; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
uri-template==1.3.0
urllib3==2.2.3; python_version >= '3.8'
watchdog==6.0.0; platform_system != 'Darwin'
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0; python_version >= '3.8'
werkzeug==3.1.3; python_version >= '3.9'
widgetsnbextension==4.0.13; python_version >= '3.7'
